Refactor
Kodlar parçalara ayrıldı
